Cleaned up code, with more understandable serialization (export method).
Sets center of viewport as the center of gravity at all times. This way we don't have the whole system moving up or down in the view.

Web Worker created only when simulation ran. All the serialized data is sent to the worker, and the data is sent back and the app is updated. Changes from the application requires that the worker first stops then, restarts with the updated data.

Better defined defaults.

Normalize method which fits all the objects in the viewport.